moved LabelAnnotation to different package
